requirements.txt
allure-pytest==2.13.1
allure-python-commons==2.13.1
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
et-xmlfile==1.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
jsonpath==0.82
numpy==1.24.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pip==23.0.1
platformdirs==3.1.0
pluggy==1.0.0
pycparser==2.21
PySocks==1.7.1
pytest-rerunfailures==11.1.2
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
setuptools==57.4.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomli==2.0.1
trio-websocket==0.9.2
trio==0.22.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.20.0
virtualenvwrapper-win==1.2.7
wsproto==1.2.0
